ProgressView and UIKit+extensions
