+++ QuizApp/QuizApp/UI/Common/UIKit+Extensions.swift
@@ -1,0 +1,3 @@
+import UIKit
+
+extension UIColor: Identifiable {}
