+++ QuizApp/QuizApp/UI/Screens/SolvingQuiz/SolvingQuizView.swift
@@ -25,29 +25,3 @@ struct SolvingQuizPreview: PreviewProvider {
     }
 
 }
-
-struct ProgressView: View {
-
-    let progressText: String
-    let progressColors: [UIColor]
-
-    var body: some View {
-        VStack {
-            Text(progressText)
-                .font(.heading5)
-                .foregroundColor(.white)
-                .maxWidth(alignment: .leading)
-
-            HStack(spacing: 5) {
-                ForEach(progressColors) {color in
-                    Rectangle()
-                        .foregroundColor(Color(color))
-                        .frame(height: 2)
-                        .cornerRadius(2)
-                }
-            }
-        }
-    }
-}
-
-extension UIColor: Identifiable {}
